+++ messages/src/main/scala/me/breidenbach/scatena/messages/SequenceUnavailableMessage.scala
@@ -6,7 +6,8 @@ import java.nio.ByteBuffer
   * @author Kevin Breidenbach 
   *         Date: 10/11/16.
   */
-case class SequenceUnavailableMessage (var startSequence: Int, var endSequence: Int) extends Message with SequenceIntervalMessage {
+case class SequenceUnavailableMessage (var startSequence: Long, var endSequence: Long)
+  extends Message with SequenceIntervalMessage {
   override protected def serializeObject(): Array[Byte] = {
     convertToByteArray(startSequence, endSequence)
   }
